tests/wpt/web-platform-tests/html/browsers/history/the-location-interface/location-prototype-setting-same-origin.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>[[SetPrototypeOf]] on a location object should not allow changing its value: same-origin</title>
<link rel="help" href="http://html.spec.whatwg.org/multipage/#location-setprototypeof">
<link rel="author" title="Domenic Denicola" href="mailto:d@domenic.me">

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="/common/test-setting-immutable-prototype.js"></script>

<script>
"use strict";

const origProto = Object.getPrototypeOf(location);

test(() => {
  assert_not_equals(origProto, null);
}, "Same-origin prerequisite check: the original prototype is accessible");

testSettingImmutablePrototype("Same-origin", location, origProto, { isSameOriginDomain: true });
</script>
